LPS-178945 SF move to proper package
